requirements.txt
ConfigSpace~=0.4.19
openbox~=0.7.9
gym~=0.17.3
keras-rl2
sql_metadata
numpy~=1.19.5
rl~=3.0
keras~=2.6.0
mabwiser~=2.4.1
scipy~=1.5.4
tqdm~=4.60.0
matplotlib~=3.3.4
sqlparse~=0.4.1
mysql-connector
pyrfr
colorama
pymysql
jpype1